requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
dj-database-url==3.0.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
Pillow==11.3.0
supabase==2.18.1
requests==2.32.4
gunicorn==23.0.0
whitenoise==6.8.0
